Add ability to turn off DDO with a message #2 (#3)
A new admin section was added that allows a DDO admin to stop accepting
DDO submissions as well as set a message that will be displayed in place
of the drop-off link. These settings are stored as properties on the /ddo
realm.

Some code cleanup was done in areas where changes were made.