Make DDO compatible with out-of-the-box Sakai (#8)
This removes the UDayton specific course management dependencies and
makes DDO compatible with Sakai 11.1. There is a version upgrade in
Sakai 11.2 that causes the build to fail. Changes will need to be made
for DDO to use the newer Sakai versions.
The DDO pack module was removed and combined with impl.